requirements.txt
# Core dependencies
streamlit
python-dotenv
pandas
uuid

# Gemini API
google-generativeai==0.8.4

# LangChain for prompt handling (used with older implementation or hybrids)
langchain-core
langchain-community

# PDF and DOCX parsing
PyPDF2
python-docx

# ChromaDB for vector store (for portfolio querying)
chromadb

# Optional but helpful for API response formatting
tqdm

# Web scraping
beautifulsoup4
requests
